@@ -27,18 +27,18 @@
         "bin/typo3-php-datasets"
     ],
     "require": {
-        "php": "^7.2 || ^7.3 || ^7.4 || ^8.0 || ^8.1 || ^8.2 || ^8.3 || ^8.4",
+        "php": "^8.2 || ^8.3 || ^8.4",
         "composer-runtime-api": "^2.2",
-        "doctrine/dbal": "^2.13 || ^3.6 || ^4.0 || 4.0.0-RC2",
-        "symfony/console": "^5.4 || ^6.2 || ^7.0",
-        "typo3/cms-core": "^10.4 || ^11.5 || ^12.4 || ^13.0"
+        "doctrine/dbal": "^4.3.3",
+        "symfony/console": "^7.2",
+        "typo3/cms-core": "^13.4.19"
     },
     "require-dev": {
         "friendsofphp/php-cs-fixer": "^3.4",
         "phpstan/phpstan": "^1.10",
         "phpstan/phpstan-phpunit": "^1.3",
-        "phpunit/phpunit": "^8.5 || ^9.6 || ^10.0",
-        "typo3/testing-framework": "^6.16 || ^7.0 || ^8.0"
+        "phpunit/phpunit": "^11.5 || ^12.2",
+        "typo3/testing-framework": "^9.2"
     },
     "extra": {
         "typo3/cms": {
